requirements.txt
# Core Flask web framework - ESSENTIAL for Vercel
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Lightweight visualization - ESSENTIAL (Graph Objects only)
plotly==5.17.0

# Supabase - ESSENTIAL (file storage only)
supabase>=2.18.0

# HTTP requests - ESSENTIAL
httpx>=0.26.0
requests>=2.31.0

# Basic utilities - ESSENTIAL
python-dotenv>=1.0.0
python-dateutil>=2.8.2
